+++ CHANGES.md
@@ -3,6 +3,11 @@
 
 Version 1.18 [????.??.??]
 -------------------------
+* Support GHC 9.12.
+* Add further support for embedded types in terms. The `DExp` type now has a
+  `DForallE` data constructor (mirroring `ForallE` and `ForallVisE` in
+  `template-haskell`) and a `DConstrainedE` data constructor (mirroring
+  `ConstrainedE` in `template-haskell`).
 * The `DLamE` and `DCaseE` data constructors (as well as the related
   `mkDLamEFromDPats` function) are now deprecated in favor of the new
   `DLamCasesE` data constructor. `DLamE`, `DCaseE`, and `mkDLamEFromDPats` will